app/requirements.txt
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2-binary==2.9.2
boto3==1.21.8
fleep==1.0.1
